feat(columns): created hook for useSortableAttributes
