+++ next-tavla/src/Admin/scenarios/Boards/hooks/useSortableAttributes.ts
@@ -1,0 +1,38 @@
+import { CSSProperties } from 'react'
+import { useSortable } from '@dnd-kit/sortable'
+import { CSS } from '@dnd-kit/utilities'
+
+export function useSortableColumnAttributes(column: string) {
+    const { attributes, listeners, transform, transition, active, setNodeRef } =
+        useSortable({ id: column })
+
+    const otherColumnActive = active && active.id !== column
+    const thisColumnActive = active && active.id === column
+
+    const activeStyle =
+        thisColumnActive &&
+        ({
+            backgroundColor: 'var(--main-background-color)',
+        } as CSSProperties)
+
+    const style = {
+        transform: CSS.Translate.toString(transform),
+        transition: transition,
+        zIndex: thisColumnActive ? 10 : 0,
+        opacity: otherColumnActive ? 0.5 : 1,
+        ...activeStyle,
+    }
+
+    return {
+        style,
+        activeStyle,
+        attributes,
+        listeners,
+        transform,
+        transition,
+        active,
+        setNodeRef,
+        thisColumnActive,
+        otherColumnActive,
+    }
+}
